Abstrat log in to component, fixes for new button styles, beginning of profile switcher
